# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "herzen"
  },
  "paths": {
    "/api/v1/login/access-token": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "Login Access Token",
        "description": "OAuth2 compatible token login, get an access token for future requests",
        "operationId": "login-login_access_token",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login-login_access_token"
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "grant_type": "string",
                    "username": "string",
                    "password": "string",
                    "scope": "",
                    "client_id": "string",
                    "client_secret": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "access_token": "string",
                      "token_type": "bearer"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/login/test-token": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "Test Token",
        "description": "Test access token",
        "operationId": "login-test_token",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/reset-password/": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "Reset Password",
        "description": "Reset password",
        "operationId": "login-reset_password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/students/{student_id}/standards-results/": {
      "get": {
        "tags": [
          "standards-results"
        ],
        "summary": "List Standards Results",
        "description": "Retrieve standard results for a given student.",
        "operationId": "standards-results-list_standards_results",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardsResultsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "standards-results"
        ],
        "summary": "Create Standard Result",
        "description": "Create new standard result.",
        "operationId": "standards-results-create_standard_result",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardResultCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/students/{student_id}/standards-results/{standard_result_id}": {
      "patch": {
        "tags": [
          "standards-results"
        ],
        "summary": "Update Standard Result",
        "description": "Update a standard result.",
        "operationId": "standards-results-update_standard_result",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          },
          {
            "name": "standard_result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Standard Result Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandardResultUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "standards-results"
        ],
        "summary": "Delete Standard Result",
        "description": "Delete a standard result.",
        "operationId": "standards-results-delete_standard_result",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          },
          {
            "name": "standard_result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Standard Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/api/v1/login/access-token",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
      "Body_login-login_access_token": {
        "type": "object",
        "title": "Body_login-login_access_token",
        "properties": {
          "grant_type": {
            "title": "Grant Type",
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ]
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "title": "Client Id",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_secret": {
            "title": "Client Secret",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "Token": {
        "type": "object",
        "title": "Token",
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "required": [
          "access_token"
        ]
      },
      "HTTPValidationError": {
        "type": "object",
        "title": "HTTPValidationError",
        "properties": {
          "detail": {
            "type": "array",
            "title": "Detail",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "title": "ValidationError",
        "properties": {
          "loc": {
            "type": "array",
            "title": "Location",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ]
      },
      "UserOut": {
        "type": "object",
        "title": "UserOut",
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "default": false
          },
          "full_name": {
            "title": "Full Name",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "required": [
          "email",
          "id"
        ]
      },
      "Message": {
        "type": "object",
        "title": "Message",
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "required": [
          "message"
        ]
      },
      "NewPassword": {
        "type": "object",
        "title": "NewPassword",
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          }
        },
        "required": [
          "token",
          "new_password"
        ]
      },
...
      "StandardsResultsOut": {
        "type": "object",
        "title": "StandardsResultsOut",
        "properties": {
          "data": {
            "title": "Data",
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StandardResultOut"
                }
              },
              {
                "type": "null"
              }
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "required": [
          "count"
        ]
      },
      "StandardResultOut": {
        "type": "object",
        "title": "StandardResultOut",
        "properties": {
          "result": {
            "type": "integer",
            "title": "Result"
          },
          "semester": {
            "type": "integer",
            "title": "Semester"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "standard_id": {
            "title": "Standard Id",
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "standard": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Standard"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "result",
          "semester",
          "id"
        ]
      },
      "Standard": {
        "type": "object",
        "title": "Standard",
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "required": [
          "name"
        ]
      },
      "StandardResultCreate": {
        "type": "object",
        "title": "StandardResultCreate",
        "properties": {
          "result": {
            "type": "integer",
            "title": "Result"
          },
          "semester": {
            "type": "integer",
            "title": "Semester"
          },
          "standard_id": {
            "type": "integer",
            "title": "Standard Id"
          }
        },
        "required": [
          "result",
          "semester",
          "standard_id"
        ]
      },
      "StandardResultUpdate": {
        "type": "object",
        "title": "StandardResultUpdate",
        "properties": {
          "result": {
            "type": "integer",
            "title": "Result"
          }
        },
        "required": [
          "result"
        ]
      },
...
    }
  }
}